# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 梗星球 MemeChatAI REST API
  version: 1.0.0
  description: |
    梗星球（MemeChatAI）REST API 接口契约，OpenAPI 3.1.0 机器可读版本。

    - 字段命名严格对齐 `db/schema.sql` v1.0.0（snake_case）。
    - 统一响应包装 `{code, data, message, request_id}`。
    - 鉴权：JWT Bearer（除标注为无需鉴权的接口）。
    - 异步 AI 任务：202 + job_id + 轮询/WebSocket。
    - M1 P0 接口对应 M1 Sprint（2026-07-08 ~ 2026-08-04）必交付范围；
      M2 接口对应第 2 个月交付；v1.5+ 接口在表格中列出但本文件给最小 schema。
    - 人类可读详设见 `docs/API-Spec.md`。
  contact:
    name: 后端架构 / API 设计
  license:
    name: Proprietary
  x-memestar-doc-version: v1.0.0
  x-memestar-related:
    prd: docs/PRD.md v1.1.0
    tech: docs/TechnicalDesign.md v1.1.0
    m1-plan: docs/M1-Sprint-Plan.md v1.0.0
    schema: docs/db/schema.sql v1.0.0
servers:
  - url: https://api.dev.memestar.cn/api/v1
    description: dev 本地开发联调
  - url: https://api.staging.memestar.cn/api/v1
    description: staging 灰度前联调 + 邀请制灰度
  - url: https://api.memestar.cn/api/v1
    description: prod 生产（ICP 备案后启用）
...
security:
  - BearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [User]
      summary: 我的资料
      description: 返回当前登录用户的完整资料，含 level/meme_power/energy_balance 等。
      operationId: getMe
      x-memestar-version: M1
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [User]
      summary: 编辑资料
      description: |
        部分字段可选更新。昵称 30 天改 1 次（由 user_profiles.nickname_changed_at 控制）。
        昵称走敏感词 DFA + 阿里云审核。
      operationId: updateMe
      x-memestar-version: M1
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserProfile'
        '400':
          description: 昵称违规或未到改昵称时间
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /memes:
    post:
      tags: [Meme]
      summary: 发布梗卡
      description: |
        从造梗会话/视频包装为梗卡，入审核队列。
        - status 流转：draft → pending_audit → published / manual_review / rejected / offline
        - AI 生成内容强制带 watermarked=true 与 is_ai_generated=true（合规）
      operationId: publishMeme
      x-memestar-version: M1
      x-memestar-rate-limit: 30/hour/user
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemePublishRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemeCard'
        '409':
          description: creation 已发布过
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                already:
                  value: { code: 3002, data: null, message: "creation 已发布过", request_id: "..." }
  /memes/feed:
    get:
      tags: [Meme, Recommend]
      summary: 推荐 feed（热度召回 v1）
      description: |
        游客可读。M1 走热度召回 v1（Redis ZSet + 新品 Top 混合 + 多样性重排）；
        M2 升级为双塔召回 + LightGBM 排序（走 /recommend/feed）。
        游标分页。
      operationId: getMemeFeed
      security: []
      x-memestar-version: M1
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
        - name: feed_type
          in: query
          schema:
            type: string
            enum: [hot, new, personalized, god]
            default: hot
        - name: legion_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Paged'
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: '#/components/schemas/MemeCardSummary'
                              next_cursor:
                                type: string
                                nullable: true
  /memes/{id}:
    get:
      tags: [Meme]
      summary: 梗卡详情
      description: 游客可读 published 状态梗卡。返回完整字段含作者摘要、视频、我的评分/收藏。
      operationId: getMeme
      security: []
      x-memestar-version: M1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemeDetail'
        '404':
          description: 梗卡不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                not_found:
                  value: { code: 3001, data: null, message: "梗卡不存在", request_id: "..." }
    delete:
      tags: [Meme]
      summary: 删除梗卡（软删除）
      description: 作者或管理员可删，软删除（deleted_at）。
      operationId: deleteMeme
      x-memestar-version: M1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功（无返回体）
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================
  # Rating
  # ============================================================
...
  /memes/{id}/comments:
    post:
      tags: [Rating]
      summary: 评论 / 回复
      description: 走敏感词 DFA + 可疑走阿里云。parent_id 楼中楼（≤3 层）。
      operationId: createComment
      x-memestar-version: M1
      x-memestar-rate-limit: 30/min/user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreateRequest'
      responses:
        '201':
          description: 评论成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Comment'
        '400':
          description: 敏感词命中 / 楼层过深
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                sensitive:
                  value: { code: 4003, data: null, message: "评论含敏感词", request_id: "..." }
                too_deep:
                  value: { code: 4004, data: null, message: "评论楼层过深", request_id: "..." }
    get:
      tags: [Rating]
      summary: 评论列表
      operationId: listComments
      x-memestar-version: M1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: sort
          in: query
          schema:
            type: string
            default: "-like_count,-created_at"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Paged'
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Comment'

  # ============================================================
  # Legion
  # ============================================================
  /legions:
    post:
      tags: [Legion]
      summary: 创建军团
      description: |
        - 创建者需 Lv.3+
        - 名称唯一（citext 不区分大小写），3-12 字，禁止敏感词与冒充官方
        - 创建时人审（status=active 但创建后排队人审，违规可 retro-ban）
        - 创建者自动成为 leader 并占用一个军团名额（≤3）
      operationId: createLegion
      x-memestar-version: M2
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LegionCreateRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Legion'
        '403':
          description: 等级不足 / 军团数上限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                max_joined:
                  value: { code: 5003, data: null, message: "已加入军团上限", request_id: "..." }
        '409':
          description: 名称重复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                dup:
                  value: { code: 5001, data: null, message: "军团名重复", request_id: "..." }
    get:
      tags: [Legion]
      summary: 军团广场列表
      operationId: listLegions
      x-memestar-version: M2
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: filter
          in: query
          schema:
            type: string
          description: "如 theme_tags:contains:抽象"
        - name: sort
          in: query
          schema:
            type: string
            default: "-activity_score"
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Paged'
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Legion'
  /legions/{id}:
    get:
      tags: [Legion]
      summary: 军团详情
      description: 含基础字段 + 成员榜 Top10 + 今日战报 + PK 战绩摘要。
      operationId: getLegion
      x-memestar-version: M2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LegionDetail'
  /legions/{id}/join:
    post:
      tags: [Legion]
      summary: 加入军团
      description: |
        - join_mode=public 立即加入
        - join_mode=approval 创建 join_request，等待队长审批（status=pending）
        - 用户已加入军团数 ≤ 3
      operationId: joinLegion
      x-memestar-version: M2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  maxLength: 100
      responses:
        '200':
          description: 加入成功 / 待审批
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LegionJoinResult'
        '403':
          description: 已加入军团上限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                max_joined:
                  value: { code: 5003, data: null, message: "已加入军团上限", request_id: "..." }

  # ============================================================
  # PK
  # ============================================================
...
  /chat/rooms/{id}/messages:
    get:
      tags: [Chat]
      summary: 消息列表
      description: 游标分页，30 天前数据归档不可查。
      operationId: listMessages
      x-memestar-version: M2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: cursor
          in: query
          schema:
            type: string
        - name: page_size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Paged'
                          - type: object
                            properties:
                              list:
                                type: array
                                items:
                                  $ref: '#/components/schemas/Message'
                              next_cursor:
                                type: string
                                nullable: true
    post:
      tags: [Chat]
      summary: 发送消息
      description: |
        - msg_type: text/image/meme/voice/system
        - extra 支持 ref_message_id（引用）、mentions（@）、ref_meme_id（梗卡分享）
        - 新成员前 24h 限 3 条/min
      operationId: sendMessage
      x-memestar-version: M2
      x-memestar-rate-limit: new 3/min, old 30/min
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageSendRequest'
      responses:
        '201':
          description: 发送成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Message'

  # ============================================================
  # Recommend
  # ============================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer Token。Header 格式：`Authorization: Bearer <access_token>`。
        access_token 有效期 2h，refresh_token 30d。

  # ============================================================
  # Parameters
  # ============================================================
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 50
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: 写操作幂等键，24h 内相同 key 返回首次响应。
...
  responses:
    Unauthorized:
      description: 未认证 / token 失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            invalid:
              value: { code: 1001, data: null, message: "token 无效", request_id: "..." }
            expired:
              value: { code: 1002, data: null, message: "token 过期", request_id: "..." }
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            forbidden:
              value: { code: 1003, data: null, message: "无权限", request_id: "..." }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            not_found:
              value: { code: 3001, data: null, message: "资源不存在", request_id: "..." }
...
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: 0=成功，非 0=业务错误码
          example: 0
        data:
          description: 业务数据，错误时为 null
        message:
          type: string
          example: ok
        request_id:
          type: string
          description: 32 位 ULID，链路追踪
          example: 01HQK8X9J2RT7N1V4Y6Z3B5M6P
    Paged:
      type: object
      required: [list, total, page, page_size, has_more]
      properties:
        list:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        has_more:
          type: boolean
...
    UserProfile:
      type: object
      description: 当前用户资料，对齐 users + user_profiles 表
      properties:
        user_id:
          type: string
          format: uuid
        nickname:
          type: string
        avatar_url:
          type: string
          nullable: true
        gender:
          type: string
          enum: [male, female, other, unknown]
        birthday:
          type: string
          format: date
          nullable: true
        bio:
          type: string
          nullable: true
        level:
          type: integer
        meme_power:
          type: integer
          description: 梗力值
        defense_value:
          type: integer
          description: 破防值
        energy_balance:
          type: integer
          description: 梗能量余额
        legion_count:
          type: integer
          description: 已加入军团数（≤3）
        is_pro:
          type: boolean
        interest_tags:
          type: array
          items:
            type: string
        badges:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, banned, teen_mode, deleted]
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        nickname:
          type: string
          maxLength: 32
        avatar_url:
          type: string
          maxLength: 512
        gender:
          type: string
          enum: [male, female, other, unknown]
        bio:
          type: string
          maxLength: 140
        birthday:
          type: string
          format: date
...
    MemePublishRequest:
      type: object
      required: [type, title, is_ai_generated]
      properties:
        creation_id:
          type: string
          format: uuid
          nullable: true
        video_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [text, image, video]
        title:
          type: string
          maxLength: 140
        tags:
          type: array
          items:
            type: string
          maxItems: 5
        legion_id:
          type: string
          format: uuid
          nullable: true
        cover_url:
          type: string
          nullable: true
        is_ai_generated:
          type: boolean
    MemeCard:
      type: object
      description: 对齐 meme_cards 表
      properties:
        meme_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        creation_id:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [text, image, video]
        cover_url:
          type: string
          nullable: true
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        legion_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [draft, pending_audit, published, manual_review, rejected, offline]
        god_trash_status:
          type: string
          enum: [pending, god, trash]
        is_ai_generated:
          type: boolean
        watermarked:
          type: boolean
        created_at:
          type: string
          format: date-time
    MemeCardSummary:
      type: object
      description: feed 列表项
      properties:
        meme_id:
          type: string
          format: uuid
        author_id:
          type: string
          format: uuid
        author_nickname:
          type: string
        author_avatar_url:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, video]
        cover_url:
          type: string
          nullable: true
        title:
          type: string
        tags:
          type: array
          items:
            type: string
        legion_id:
          type: string
          format: uuid
          nullable: true
        score_avg:
          type: number
          format: float
        score_count:
          type: integer
        comment_count:
          type: integer
        share_count:
          type: integer
        hot_score:
          type: number
          format: float
        god_trash_status:
          type: string
          enum: [pending, god, trash]
        is_ai_generated:
          type: boolean
        published_at:
          type: string
          format: date-time
    MemeDetail:
      type: object
      allOf:
        - $ref: '#/components/schemas/MemeCardSummary'
        - type: object
          properties:
            author:
              $ref: '#/components/schemas/UserSummary'
            legion:
              type: object
              nullable: true
              properties:
                legion_id:
                  type: string
                  format: uuid
                name:
                  type: string
                avatar_url:
                  type: string
            favorite_count:
              type: integer
            view_count:
              type: integer
            completion_rate:
              type: number
              format: float
            video:
              $ref: '#/components/schemas/VideoStatus'
              nullable: true
            my_rating:
              type: integer
              nullable: true
              description: 我的评分（1-5），未评分为 null
            my_favorite:
              type: boolean
            published_at:
              type: string
              format: date-time
              nullable: true
            created_at:
              type: string
              format: date-time

    # ---------- Rating / Comment ----------
...
    CommentCreateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 500
        parent_id:
          type: string
          format: uuid
          nullable: true
        ref_meme_id:
          type: string
          format: uuid
          nullable: true
          description: 引用梗卡（造梗接龙）
        is_meme_card:
          type: boolean
          description: 评论本身是否为新梗卡
    Comment:
      type: object
      description: 对齐 comments 表
      properties:
        comment_id:
          type: string
          format: uuid
        meme_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        content:
          type: string
        like_count:
          type: integer
        is_god_comment:
          type: boolean
        status:
          type: string
          enum: [published, hidden, deleted]
        created_at:
          type: string
          format: date-time

    # ---------- Legion ----------
    LegionCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 12
        slogan:
          type: string
          maxLength: 140
        avatar_url:
          type: string
        theme_tags:
          type: array
          items:
            type: string
        join_mode:
          type: string
          enum: [public, approval]
          default: approval
    Legion:
      type: object
      description: 对齐 legions 表
      properties:
        legion_id:
          type: string
          format: uuid
        name:
          type: string
        slogan:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        theme_tags:
          type: array
          items:
            type: string
        leader_id:
          type: string
          format: uuid
        level:
          type: integer
        activity_score:
          type: integer
        member_count:
          type: integer
        member_cap:
          type: integer
        join_mode:
          type: string
          enum: [public, approval]
        badges:
          type: array
          items:
            type: string
        pk_wins:
          type: integer
        pk_losses:
          type: integer
        status:
          type: string
          enum: [active, frozen, dissolved]
        created_at:
          type: string
          format: date-time
    LegionDetail:
      allOf:
        - $ref: '#/components/schemas/Legion'
        - type: object
          properties:
            top_members:
              type: array
              items:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  nickname:
                    type: string
                  avatar_url:
                    type: string
                  contribution:
                    type: integer
                  role:
                    type: string
            today_report:
              type: object
              properties:
                new_members:
                  type: integer
                new_memes:
                  type: integer
                pk_progress:
                  type: string
    LegionJoinResult:
      type: object
      properties:
        membership_id:
          type: string
          format: uuid
        legion_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [leader, vice_leader, member]
        contribution:
          type: integer
        joined_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [joined, pending]

    # ---------- PK ----------
...